+++ src/Report/AbstractReport2TableGUI.php
@@ -326,10 +326,6 @@ abstract class AbstractReport2TableGUI extends AbstractReportTableGUI
 
         $check_agreement = false;
 
-        if (count($additional_fields) > 0) {
-            unset($additional_fields["condition_passed"]);
-        }
-
         $tr_data = ilTrQuery::getUserDataForObject($this->ref_id, ilUtil::stripSlashes($this->getOrderField()), ilUtil::stripSlashes($this->getOrderDirection()),
             ilUtil::stripSlashes($this->getOffset()), ilUtil::stripSlashes($this->getLimit()), $filter, $additional_fields, $check_agreement, $this->user_fields);